requirements.txt
asgiref==3.6.0
Django==4.1.5
Pillow==9.4.0
sqlparse==0.4.3
django-bootstrap5==22.2 
django-unfold==0.21.1